Test 23

+++ src/Homeworks/Test23/Individual.java
@@ -15,14 +15,14 @@ public class Individual extends Client {
     @Override
     void refillAccount() {
         System.out.print("Введите сумму пополнения: ");
-        int refillSum = new Scanner(System.in).nextInt();
+        double refillSum = new Scanner(System.in).nextDouble();
         accountSum += refillSum * (1 - refillFeePercentage * 0.01);
     }
 
     @Override
     void withdrawalFromAccount() {
         System.out.print("Введите сумму снятия: ");
-        int withdrawalSum = new Scanner(System.in).nextInt();
+        double withdrawalSum = new Scanner(System.in).nextDouble();
         accountSum -= withdrawalSum * (1 - withdrawalFeePercentage * 0.01);
     }
